add subscribeplan method

+++ database/migrations/2022_06_24_151646_create_cover_plans_table.php
@@ -17,7 +17,7 @@ class CreateCoverPlansTable extends Migration
              $table->id();
             $table->string('cover_type');
             $table->string('cover_flat');
-            $table->integer('cover_price');
+            $table->double('cover_price');
             $table->timestamps();
         });
     }
